Conflicts:
	src/main/java/com/umeet/umeet/repositories/ChannelRepository.java
	src/main/java/com/umeet/umeet/repositories/MessageRepository.java

+++ src/main/java/com/umeet/umeet/repositories/ProfileRepository.java
@@ -5,6 +5,6 @@ import com.umeet.umeet.entities.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface ProfileRepository extends JpaRepository<User, Integer>{
-    
+public interface ProfileRepository extends JpaRepository<User, Long>{
+
 }
